// apps/desktop/src/types/editor.ts
export interface Pagination {
  page: number;
  page_size: number;
  total_rows: number | null;
  has_more: boolean;
}